Checkpoint: More moves of cf-specific code
