+++ src/frontend/packages/cloud-foundry/src/cf-entity-factory.ts
@@ -4,7 +4,7 @@ import { EntitySchema } from '../../store/src/helpers/entity-schema';
 import { APIResource } from '../../store/src/types/api.types';
 import { CfUser, CfUserRoleParams, OrgUserRoleNames, SpaceUserRoleNames } from './store/types/user.types';
 import { CF_ENDPOINT_TYPE } from '../cf-types';
-import { getAPIResourceGuid } from './selectors/api.selectors';
+import { getAPIResourceGuid } from './store/selectors/api.selectors';
 
 export const applicationEntityType = 'application';
 export const stackEntityType = 'stack';
