+++ src/frontend/packages/cloud-foundry/src/store/selectors/cf-current-user-role.selectors.ts
@@ -1,39 +1,25 @@
 import { compose } from '@ngrx/store';
 
 import {
+  PermissionStrings,
+  PermissionValues,
+  ScopeStrings,
+} from '../../../../core/src/core/current-user-permissions.config';
+import {
+  selectCurrentUserCFGlobalHasScopes,
+  selectCurrentUserRequestState,
+  selectCurrentUserRolesState,
+} from '../../../../store/src/selectors/current-user-role.selectors';
+import { ICurrentUserRolesState } from '../../../../store/src/types/current-user-roles.types';
+import {
   IAllCfRolesState,
   ICfRolesState,
   IGlobalRolesState,
   IOrgsRoleState,
   ISpacesRoleState,
-  RoleEntities,
-} from '../../../../cloud-foundry/src/store/types/cf-current-user-roles.types';
-import {
-  PermissionStrings,
-  PermissionValues,
-  ScopeStrings,
-} from '../../../../core/src/core/current-user-permissions.config';
-import { CurrentUserRolesAppState } from '../../app-state';
-import { ICurrentUserRolesState, IStratosRolesState } from '../../types/current-user-roles.types';
+} from '../types/cf-current-user-roles.types';
 
-
-type Omit<T, K extends keyof T> = Pick<T, Exclude<keyof T, K>>;
-
-export const selectCurrentUserRolesState = (state: CurrentUserRolesAppState) => state.currentUserRoles;
-
-export const selectCurrentUserStratosRolesState = (state: ICurrentUserRolesState) => state.internal;
-
-export const selectCurrentUserStratosRoles = (role: PermissionValues) => (state: Omit<IStratosRolesState, 'scopes'>) => {
-  // Note - should not cover `scopes`
-  return state[role] || false;
-};
-
-export const selectEntityWithRole = (role: PermissionStrings, type: RoleEntities) => (state: ICfRolesState) => {
-  const entityType = state[type];
-  return Object.keys(entityType).filter(entity => entityType[entity][role]);
-};
-
-export const selectSpaceWithRoleFromOrg = (role: PermissionStrings, orgId: string) => (state: ICfRolesState) => {
+const selectSpaceWithRoleFromOrg = (role: PermissionStrings, orgId: string) => (state: ICfRolesState) => {
   if (!state) {
     return 'all';
   }
@@ -56,58 +42,19 @@ export const selectCurrentUserCFRolesState = (state: ICurrentUserRolesState) => 
 export const selectCurrentUserCFEndpointRolesState = (endpointGuid: string) =>
   (state: IAllCfRolesState) => state ? state[endpointGuid] : null;
 
-export const selectCurrentUserRequestState = (state: ICurrentUserRolesState | ICfRolesState) => state.state;
-
 export const selectCurrentUserCFGlobalRolesStates = (state: ICfRolesState) => state ? state.global : null;
 export const selectCurrentUserCFGlobalRolesState = (role: PermissionValues) => (state: IGlobalRolesState) => state[role] || false;
 export const selectCurrentUserCFOrgsRolesState = (state: ICfRolesState) => state.organizations;
 export const selectCurrentUserCFSpacesRolesState = (state: ICfRolesState) => state.spaces;
-export const selectCurrentUserCFGlobalHasScopes = (scope: ScopeStrings) => (scopes: ScopeStrings[]) => scopes.includes(scope);
 export const selectCurrentUserCFGlobalScopesState = (state: IGlobalRolesState) => state.scopes;
-export const selectCurrentUserCFStratosScopesState = (state: IStratosRolesState) => state.scopes;
 
 export const selectCurrentUserCFSpaceRolesState = (spaceId: string) => (state: ISpacesRoleState) => state[spaceId];
 export const selectCurrentUserCFOrgRolesState = (orgId: string) => (state: IOrgsRoleState) => state[orgId];
-
-// Top level stratos endpoint role objects
-// ============================
-export const getCurrentUserStratosRolesState = compose(
-  selectCurrentUserStratosRolesState,
-  selectCurrentUserRolesState
-);
-// ============================
-
-// Top level stratos endpoint role objects
-// ============================
-export const getCurrentUserStratosRole = (role: PermissionValues) => compose(
-  selectCurrentUserStratosRoles(role),
-  getCurrentUserStratosRolesState
-);
-// ============================
-
-// Top level stratos endpoint scopes
-// ============================
-export const getCurrentUserStratosHasScope = (scope: ScopeStrings) => compose(
-  selectCurrentUserCFGlobalHasScopes(scope),
-  selectCurrentUserCFStratosScopesState,
-  getCurrentUserStratosRolesState
-);
-// ============================
-
-
 
 // Top level cf endpoint role objects
 // ============================
 export const getCurrentUserCFRolesState = compose(
   selectCurrentUserCFRolesState,
-  selectCurrentUserRolesState
-);
-// ============================
-
-// Top level request state
-// ============================
-export const getCurrentUserRequestState = compose(
-  selectCurrentUserRequestState,
   selectCurrentUserRolesState
 );
 // ============================
@@ -183,7 +130,6 @@ export const getCurrentUserCFSpaceRolesState = (endpointGuid: string, spaceId: s
   getCurrentUserCFSpacesRolesState(endpointGuid)
 );
 // ============================
-
 
 // Specific org roles
 // ============================
