+++ src/frontend/packages/core/src/features/applications/application.service.ts
@@ -27,7 +27,7 @@ import {
   spaceEntityType,
   stackEntityType,
 } from '../../../../cloud-foundry/src/cf-entity-factory';
-import { selectCfEntity } from '../../../../cloud-foundry/src/selectors/api.selectors';
+import { selectCfEntity } from '../../../../cloud-foundry/src/store/selectors/api.selectors';
 import { createEntityRelationKey } from '../../../../store/src/helpers/entity-relations/entity-relations.types';
 import { ActionState, rootUpdatingKey } from '../../../../store/src/reducers/api-request-reducer/types';
 import {
