+++ src/frontend/packages/core/src/features/applications/deploy-application/deploy-application-options-step/deploy-application-options-step.component.ts
@@ -11,7 +11,7 @@ import { FetchAllDomains } from '../../../../../../cloud-foundry/src/actions/dom
 import { GetAllStacks } from '../../../../../../cloud-foundry/src/actions/stack.action';
 import { CFAppState } from '../../../../../../cloud-foundry/src/cf-app-state';
 import { getPaginationObservables } from '../../../../../../store/src/reducers/pagination-reducer/pagination-reducer.helper';
-import { selectCfDetails } from '../../../../../../cloud-foundry/src/selectors/deploy-application.selector';
+import { selectCfDetails } from '../../../../../../cloud-foundry/src/store/selectors/deploy-application.selector';
 import { APIResource } from '../../../../../../store/src/types/api.types';
 import { OverrideAppDetails } from '../../../../../../cloud-foundry/src/store/types/deploy-application.types';
 import { IDomain } from '../../../../core/cf-api.types';
