+++ src/frontend/packages/core/src/shared/components/add-service-instance/select-plan-step/select-plan-step.component.ts
@@ -29,7 +29,7 @@ import {
   SetCreateServiceInstanceServicePlan,
 } from '../../../../../../cloud-foundry/src/actions/create-service-instance.actions';
 import { CFAppState } from '../../../../../../cloud-foundry/src/cf-app-state';
-import { selectCreateServiceInstance } from '../../../../../../cloud-foundry/src/selectors/create-service-instance.selectors';
+import { selectCreateServiceInstance } from '../../../../../../cloud-foundry/src/store/selectors/create-service-instance.selectors';
 import { APIResource } from '../../../../../../store/src/types/api.types';
 import { IServicePlan } from '../../../../core/cf-api-svc.types';
 import { safeUnsubscribe } from '../../../../core/utils.service';
