+++ src/frontend/packages/core/src/shared/components/add-service-instance/specify-details-step/specify-details-step.component.ts
@@ -38,13 +38,13 @@ import {
 } from '../../../../../../cloud-foundry/src/actions/service-instances.actions';
 import { CFAppState } from '../../../../../../cloud-foundry/src/cf-app-state';
 import { serviceBindingEntityType, serviceInstancesEntityType } from '../../../../../../cloud-foundry/src/cf-entity-factory';
-import { selectCfRequestInfo, selectCfUpdateInfo } from '../../../../../../cloud-foundry/src/selectors/api.selectors';
+import { selectCfRequestInfo, selectCfUpdateInfo } from '../../../../../../cloud-foundry/src/store/selectors/api.selectors';
 import { RouterNav } from '../../../../../../store/src/actions/router.actions';
 import { getDefaultRequestState, RequestInfoState } from '../../../../../../store/src/reducers/api-request-reducer/types';
 import {
   selectCreateServiceInstance,
   selectCreateServiceInstanceSpaceGuid,
-} from '../../../../../../cloud-foundry/src/selectors/create-service-instance.selectors';
+} from '../../../../../../cloud-foundry/src/store/selectors/create-service-instance.selectors';
 import { APIResource, NormalizedResponse } from '../../../../../../store/src/types/api.types';
 import { CreateServiceInstanceState } from '../../../../../../cloud-foundry/src/store/types/create-service-instance.types';
 import { IServiceInstance, IServicePlan } from '../../../../core/cf-api-svc.types';
