+++ src/frontend/packages/core/src/shared/components/list/list-types/cf-spaces-service-instances/table-cell-service-plan/table-cell-service-plan.component.ts
@@ -8,7 +8,7 @@ import {
   servicePlanEntityType,
   userProvidedServiceInstanceEntityType,
 } from '../../../../../../../../cloud-foundry/src/cf-entity-factory';
-import { selectCfEntity } from '../../../../../../../../cloud-foundry/src/selectors/api.selectors';
+import { selectCfEntity } from '../../../../../../../../cloud-foundry/src/store/selectors/api.selectors';
 import { APIResource } from '../../../../../../../../store/src/types/api.types';
 import { IServicePlan } from '../../../../../../core/cf-api-svc.types';
 import { TableCellCustom } from '../../../list.types';
