+++ src/frontend/packages/core/src/shared/components/list/list-types/github-commits/github-commits-list-config-deploy.service.ts
@@ -8,7 +8,7 @@ import { GithubCommitsDataSource } from './github-commits-data-source';
 import { TableCellRadioComponent } from '../../list-table/table-cell-radio/table-cell-radio.component';
 import { CFAppState } from '../../../../../../../cloud-foundry/src/cf-app-state';
 import { DeployApplicationSource } from '../../../../../../../cloud-foundry/src/store/types/deploy-application.types';
-import { selectApplicationSource } from '../../../../../../../cloud-foundry/src/selectors/deploy-application.selector';
+import { selectApplicationSource } from '../../../../../../../cloud-foundry/src/store/selectors/deploy-application.selector';
 import { GitSCMType, GitSCMService } from '../../../../data-services/scm/scm.service';
 
 @Injectable()
