+++ src/frontend/packages/core/src/shared/services/cf-org-space-label.service.ts
@@ -4,7 +4,7 @@ import { filter, first, map } from 'rxjs/operators';
 
 import { CFAppState } from '../../../../cloud-foundry/src/cf-app-state';
 import { organizationEntityType, spaceEntityType } from '../../../../cloud-foundry/src/cf-entity-factory';
-import { selectCfEntity } from '../../../../cloud-foundry/src/selectors/api.selectors';
+import { selectCfEntity } from '../../../../cloud-foundry/src/store/selectors/api.selectors';
 import { endpointSchemaKey } from '../../../../store/src/helpers/entity-factory';
 import { selectEntity } from '../../../../store/src/selectors/api.selectors';
 import { APIResource } from '../../../../store/src/types/api.types';
