+++ src/frontend/packages/store/src/selectors/internal-events.selectors.ts
@@ -1,11 +1,10 @@
 import { compose, createSelector } from '@ngrx/store';
 
-import { CFAppState } from '../../../cloud-foundry/src/cf-app-state';
 import { InternalEventSeverity, InternalEventsState } from '../types/internal-events.types';
+import { InternalAppState } from '../app-state';
 
 
-// TODO: Confirm - should this be CFAppState?
-export const internalEventStateSelector = (state: CFAppState) => state.internalEvents;
+export const internalEventStateSelector = (state: InternalAppState) => state.internalEvents;
 
 export const internalEventTypeSelector = (type: string) => compose(
   (state: InternalEventsState) => state.types[type] || {},
